Refactor search components and add new arrangement component
