--- app/components/results.tsx
@@ -1,11 +1,0 @@
-export default function Results() {
-    return (
-        <main>
-            <div className="mainBox">
-                <div className="arrangement">
-                    <p>Content</p>
-                </div>
-            </div>
-        </main>
-    )
-}
