+++ app/components/search/results.tsx
@@ -1,0 +1,11 @@
+import Arrangement from "./arrangement"
+
+export default function Results() {
+    return (
+        <main className="mainBox flex-col basis-1/2 min-h-1/2">
+            <Arrangement />
+            <Arrangement />
+            <Arrangement />
+        </main>
+    )
+}
