+++ app/components/search/search.tsx
@@ -1,6 +1,6 @@
 export default function Search() {
     return (
-        <main>
+        <main className="searchBox flex justify-center items-center w-screen h-16">
             SearchBar
         </main>
     )
