+++ app/search/page.tsx
@@ -1,12 +1,12 @@
-import Results from '@/app/components/results';
-import SearchBar from '@/app/components/search';
-import Filter from '@/app/components/filter';
+import Results from '@/app/components/search/results';
+import SearchBar from '@/app/components/search/search';
+import Filter from '@/app/components/search/filter';
 
 export default function Search() {
     return (
         <main>
             <SearchBar />
-            <div className="mainContent">
+            <div className="searchContent flex flex-row justify-end items-center p-24">
                 <Results />
                 <Filter />
             </div>
